training/requirements.txt
numpy>=2.0
python-chess>=1.999
torch>=2.2
